Rename scroll-bar-hider to scrollbar-hider
